refactor: QuestionScheduler 위치 이동 #71
