+++ src/test/java/kr/co/csalgo/infrastructure/question/QuestionSchedulerTest.java
@@ -10,6 +10,7 @@ import org.mockito.Mock;
 import org.mockito.MockitoAnnotations;
 
 import kr.co.csalgo.application.problem.usecase.SendDailyQuestionMailUseCase;
+import kr.co.csalgo.presentation.scheduler.QuestionScheduler;
 
 @DisplayName("QuestionScheduler 테스트")
 class QuestionSchedulerTest {
